+++ packages/fcl/CHANGELOG.md
@@ -1,7 +1,9 @@
 ## Unreleased
 
+- 2021-11-19 -- Update `buildUser` with with correct data prop for `expiresAt`
+
 - YYYY-MM-DD **BREAKING?** -- description
-- 
+
 ## 0.0.78-alpha.8 - 2021-11-17
 
 - 2021-11-17 -- VSN `@onflow/types` 0.0.5 -> 0.0.6